correct error in post request for adduser php

+++ adduser.php
@@ -3,14 +3,13 @@ require_once 'functions.php';
 
 if (isset($_POST['userID']))
 {
-    $userID = sanitizeString($_POST['userID']);
+    $userID = (int) sanitizeString($_POST['userID']);
     $userName = sanitizeString($_POST['userName']);
     
     if (mysqli_num_rows($connect, queryMysql($connect, "SELECT * FROM members WHERE user='$userID'")))
         null;
     else {
         queryMysql($connect, "INSERT INTO user VALUES($userID, '$userName')");
-        
     }
 }
 ?>
